feat: tracking fields and code structure refactor (#11)
* chore: bump deps

* chore: remove redundant `NewGenericSyntax` mypy flag

* chore: remove unused `lib.common` module

* refactor: restructure modules

* chore: add `pika` type stubs

* chore: remove unused scripts

* refactor: clean up unused defs + data models and unify `Runner` and `Executor`

* refactor: use `importlib` for script path resolution

Similarly, `jinja` templates are now resolved using `PackageLoader`.
This is preferred as we no longer have to assume the location of `cwd`.

* feat: abstract out executor file system creation to base `Executor`

* refactor: `Sandbox` inherit from `Unsafe`

They are practically the same when it comes to setup, they are only
different during execution.

* refactor: abstract out executor result processing

* feat: save `run.sh` entrypoint to filesystem

This will help save a ton of time when it comes to debugging and program replay.

* fix: `SandboxExecutor` program process entrypoint

* fix: `SandboxExecutor` exit_code file path

* feat: only cleanup executor working directory if no exception

* feat: error if `CONTY_PATH` does not exist

This is to help separate internal runtime errors and program runtime
errors.

* fix: ack after job is done and pushed

Also fixed some errors with function currying

* refactor: central module for all data models